Use createSlice with teams
